Remove remaining public stdlib alias residue

+++ demos/guardrails/emitted.rs
@@ -1,4 +1,5 @@
+// src/main.rs
 fn main() {
     println!("guardrails guardrails demo:");
-    println!("{}", 20 as i64);
+    println!("{}", 20_i64);
 }
